pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "flask-dance>=7.1.0",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
    "anthropic>=0.57.1",
    "google-generativeai>=0.8.5",
    "requests>=2.32.4",
]
